+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - VPC Routing Table routes can now be assigned a Priority value and can advertise to the parent table's sources by using the `Advertise` toggle
 - Cloud Object Storage buckets now support Metrics Monitoring and Activity Tracking
 - Local CIDRs and Peer CIDRs are no longer required when creating a connection for a route-based VPN Gateway
+- F5 Templates with an invalid TMOS Admin Password are now shown as invalid from `/v2/vpcDeployments`
 
 ### Fixes
 
@@ -17,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - Fixed an issue where VPC address prefixes were not being updated when advanced subnets were created with Manual Address Management enabled
 - Fixed an issue preventing Terraform from being created when using an existing VPC
 - Fixed an issue causing Routing Tables to not have correct reference to an imported VPC
+- Clusters with names longer than 32 characters are now correctly shown as invalid
 - Fixed an issue where the VPN as a service template had an incorrect VPC address prefix
 
 ## 1.14.1